Make sure this actually works in a production environment.
So there's an undocumented (other than code) requirement that each
pattern behave exactly like a CommandPatterns object.  This code in its
original form stepped on that assumption.  The quick fix is to do what
I've done here, make it a sub-class of CommandPatterns.  Not sold that
this is the 100% correct way to handle this (I'd prefer more flexibility
in nappingcat.serve.ServeApp personally), but it works.